@@ -16,53 +16,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.6",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.3.6",
-    "@angular/compiler": "^4.3.6",
-    "@angular/core": "^4.3.6",
-    "@angular/forms": "^4.3.6",
-    "@angular/http": "^4.3.6",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.3.6",
-    "@angular/platform-browser-dynamic": "^4.3.6",
-    "@angular/router": "^4.3.6",
-    "angularfire2": "5.0.0-rc.3",
-    "core-js": "^2.4.1",
-    "enhanced-resolve": "^3.3.0",
-    "firebase": "^4.5.0",
-    "hammerjs": "^2.0.8",
-    "ng2-ckeditor": "^1.1.9",
-    "ng2-dnd": "^4.2.0",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "@angular/animations": "~11.2.3",
+    "@angular/cdk": "^11.2.7",
+    "@angular/common": "~11.2.3",
+    "@angular/compiler": "~11.2.3",
+    "@angular/core": "~11.2.3",
+    "@angular/fire": "^6.1.4",
+    "@angular/forms": "~11.2.3",
+    "@angular/material": "^11.2.7",
+    "@angular/platform-browser": "~11.2.3",
+    "@angular/platform-browser-dynamic": "~11.2.3",
+    "@angular/router": "~11.2.3",
+    "@types/ckeditor": "^4.9.10",
+    "core-js": "^3.11.0",
+    "firebase": "^7.0 || ^8.0",
+    "firebase-admin": "^9.6.0",
+    "firebase-functions": "^3.13.2",
+    "ng2-ckeditor": "^1.3.3",
+    "reactive-forms-typed": "^11.0.4",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.1",
-    "@angular/compiler-cli": "^4.3.6",
-    "@angular/language-service": "^4.3.6",
-    "@angular/service-worker": "^1.0.0-beta.16",
-    "@types/ckeditor": "^0.0.34",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^8.0.19",
-    "cheerio": "^1.0.0-rc.2",
-    "codelyzer": "~3.0.1",
+    "@angular-devkit/architect": ">= 0.900 < 0.1200",
+    "@angular-devkit/build-angular": "~0.1102.2",
+    "@angular/cli": "~11.2.2",
+    "@angular/compiler-cli": "~11.2.3",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
+    "cheerio": "^1.0.0-rc.6",
+    "codelyzer": "^6.0.0",
+    "firebase-tools": "^8.0.0",
     "fs": "0.0.1-security",
-    "inquirer": "^3.2.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "mkdirp": "^0.5.1",
-    "nightmare": "^2.10.0",
-    "protractor": "~5.1.2",
-    "ts-node": "^3.3.0",
-    "tslint": "^5.5.0",
-    "typescript": "^2.5.1"
+    "fuzzy": "^0.1.3",
+    "inquirer": "^8.0.0",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "mkdirp": "^1.0.4",
+    "nightmare": "^3.0.2",
+    "open": "^7.0.3",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.1.2"
   }
 }
