@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFireDatabase, AngularFireList, AngularFireObject } from 'angularfire2/database';
+import { AngularFireDatabase, AngularFireList, AngularFireObject } from '@angular/fire/database';
 import * as firebase from 'firebase/app';
-import { FirebaseApp } from 'angularfire2';
-import { GlobalService } from 'app/services/global.service';
+import { FirebaseApp } from '@angular/fire';
+import { GlobalService } from '../../../services/global.service';
 import { Router }    from '@angular/router';
-import { MdSnackBar, MdDialogRef, MdDialog } from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { DeleteDialogComponent } from '../delete-dialog/delete-dialog.component'
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'admin-products',
@@ -18,7 +19,7 @@ export class AdminProductsComponent implements OnInit {
   products: Observable<any>;
   product: AngularFireObject<any>;
   selectedOption: any;
-  dialogRef: MdDialogRef<any>;
+    dialogRef: MatDialogRef<any>;
   storageRef: any;
   currentAdmin: any;
 
@@ -27,8 +28,8 @@ export class AdminProductsComponent implements OnInit {
     public db: AngularFireDatabase,
     public globalService: GlobalService,
     public router: Router,
-    public dialog: MdDialog,
-    public snackBar: MdSnackBar
+      public dialog: MatDialog,
+    public snackBar: MatSnackBar
   ) {
     this.products = db.list('/products', ref => ref.orderByChild('rdateUpdated').limitToLast(9999)).snapshotChanges();
     this.storageRef = af.storage().ref();
