@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { GlobalService } from '../../services/global.service';
-import { AngularFireDatabase } from 'angularfire2/database';
-import { AngularFireAuth } from 'angularfire2/auth';
+import { AngularFireDatabase } from '@angular/fire/database';
+import { AngularFireAuth } from '@angular/fire/auth';
+import * as firebase from 'firebase/app';
 
 @Component({
   selector: 'cart-icon',
@@ -11,7 +12,7 @@ import { AngularFireAuth } from 'angularfire2/auth';
 })
 export class CartIconComponent implements OnInit {
   globalCart: any;
-  user: Observable<firebase.User>;
+  user: Observable<firebase.default.User>;
   cartItems = 0;
 
   constructor(public globalService: GlobalService, public afAuth: AngularFireAuth, public db: AngularFireDatabase) {
