@@ -1,33 +1,32 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { AdminComponent } from 'app/admin/admin-components/admin/admin.component';
-import { AddPostComponent } from 'app/admin/admin-components/add-post/add-post.component';
-import { AdminPostsComponent } from 'app/admin/admin-components/admin-posts/admin-posts.component';
-import { AdminAdminsComponent } from 'app/admin/admin-components/admin-admins/admin-admins.component';
-import { AdminDashboardComponent } from 'app/admin/admin-components/admin-dashboard/admin-dashboard.component';
-import { AddAdminComponent } from 'app/admin/admin-components/add-admin/add-admin.component';
-import { AdminPagesComponent } from 'app/admin/admin-components/admin-pages/admin-pages.component';
-import { AddPageComponent } from 'app/admin/admin-components/add-page/add-page.component';
-import { AdminMenusComponent } from 'app/admin/admin-components/admin-menus/admin-menus.component';
-import { AdminThemeComponent } from 'app/admin/admin-components/admin-theme/admin-theme.component';
-import { AdminProductsComponent } from 'app/admin/admin-components/admin-products/admin-products.component';
-import { AddProductComponent } from 'app/admin/admin-components/add-product/add-product.component';
-import { DeleteDialogComponent } from 'app/admin/admin-components/delete-dialog/delete-dialog.component';
-import { AdminCustomersComponent } from 'app/admin/admin-components/admin-customers/admin-customers.component';
-import { AddOrderComponent } from 'app/admin/admin-components/add-order/add-order.component';
-import { AddCustomerComponent } from 'app/admin/admin-components/add-customer/add-customer.component';
-import { AdminOrdersComponent } from 'app/admin/admin-components/admin-orders/admin-orders.component';
-import { AdminProductCategoriesComponent } from 'app/admin/admin-components/admin-product-categories/admin-product-categories.component';
-import { AddProductCategoryComponent } from 'app/admin/admin-components/add-product-category/add-product-category.component';
-import { AdminApprovalsComponent } from 'app/admin/admin-components/admin-approvals/admin-approvals.component';
-import { ApproveDialogComponent } from 'app/admin/admin-components/approve-dialog/approve-dialog.component';
-import { OrderComponent } from 'app/storefront-components/order/order.component';
-import { OrdersComponent } from 'app/storefront-components/orders/orders.component';
+import { AdminComponent } from './admin-components/admin/admin.component';
+import { AddPostComponent } from './admin-components/add-post/add-post.component';
+import { AdminPostsComponent } from './admin-components/admin-posts/admin-posts.component';
+import { AdminAdminsComponent } from './admin-components/admin-admins/admin-admins.component';
+import { AdminDashboardComponent } from './admin-components/admin-dashboard/admin-dashboard.component';
+import { AddAdminComponent } from './admin-components/add-admin/add-admin.component';
+import { AdminPagesComponent } from './admin-components/admin-pages/admin-pages.component';
+import { AddPageComponent } from './admin-components/add-page/add-page.component';
+import { AdminMenusComponent } from './admin-components/admin-menus/admin-menus.component';
+import { AdminThemeComponent } from './admin-components/admin-theme/admin-theme.component';
+import { AdminProductsComponent } from './admin-components/admin-products/admin-products.component';
+import { AddProductComponent } from './admin-components/add-product/add-product.component';
+import { DeleteDialogComponent } from './admin-components/delete-dialog/delete-dialog.component';
+import { AdminCustomersComponent } from './admin-components/admin-customers/admin-customers.component';
+import { AddOrderComponent } from './admin-components/add-order/add-order.component';
+import { AddCustomerComponent } from './admin-components/add-customer/add-customer.component';
+import { AdminOrdersComponent } from './admin-components/admin-orders/admin-orders.component';
+import { AdminProductCategoriesComponent } from './admin-components/admin-product-categories/admin-product-categories.component';
+import { AddProductCategoryComponent } from './admin-components/add-product-category/add-product-category.component';
+import { AdminApprovalsComponent } from './admin-components/admin-approvals/admin-approvals.component';
+import { ApproveDialogComponent } from './admin-components/approve-dialog/approve-dialog.component';
+import { OrderComponent } from '../storefront-components/order/order.component';
 
-import { AuthGuard } from 'app/services/auth.guard';
-import { SuperAdminGuard } from 'app/services/super-admin.guard';
-import { AdminGuard } from 'app/services/admin.guard';
+import { AuthGuard } from '../services/auth.guard';
+import { SuperAdminGuard } from '../services/super-admin.guard';
+import { AdminGuard } from '../services/admin.guard';
 
 const routes: Routes = [
   { path: '', component: AdminComponent, canActivate: [AuthGuard],
