@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Title, Meta } from '@angular/platform-browser';
 import { Router } from '@angular/router';
-import { MdSnackBar } from '@angular/material';
-import { GlobalService } from 'app/services/global.service';
-import { LocalCartService } from "app/services/localcart.service";
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { LocalCartService } from '../../services/localcart.service';
+import { GlobalService } from '../../services/global.service';
+
 
 @Component({
   selector: 'checkout-billing',
@@ -16,7 +17,7 @@ export class CheckoutBillingComponent implements OnInit {
   sameAsShipping: boolean;
 
   constructor(
-    public snackBar: MdSnackBar,
+    public snackBar: MatSnackBar,
     public router: Router,
     public globalService: GlobalService,
     public localCart: LocalCartService,
@@ -63,7 +64,7 @@ export class CheckoutBillingComponent implements OnInit {
       } else {
           let snackBarRef = this.snackBar.open('You must complete the form', 'OK!', {
             duration: 3000,
-            extraClasses: ['warn-snackbar']
+            panelClass: ['warn-snackbar']
           });
       }
   }
